genai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blockbuster==1.5.25
boto3==1.40.5
botocore==1.40.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-api==0.2.117
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.6
langgraph-prebuilt==0.6.2
langgraph-runtime-inmem==0.6.8
langgraph-sdk==0.2.0
langsmith==0.4.10
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.98.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==2.1.3
starlette==0.47.2
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
